Small update to the bug fix to handle the edge case described in PR #42.
Edge Case:
consider a create and keep request that requests 2 entangled pairs. If the first pair is successful
and the second pair fails, we will not free the memory used for the first pair, but the task will fail.
This small update fixes this, so that if one of the pairs fails it ensures that all of the quantum memories
will be freed.